Update GenericRecognizer.cs
Issue #500
I have removed the event post processing code into a separate method. This enables a call to post processing independently of the previous profile event detection.
This is useful for non-generic call recognizers. In particular, it is useful for the Koala3 call recognizer where I wanted to filter the  Koala profile events on features that are specific to the koala bellow.