Refactor methods for calculation of event statistics.
Issue #500 The calculation of event statistics was refactored in order to be able to access this functionality from within the KoalaMark3 recognizer.
Can now calculate event statistics by supplying the decibel spectrogram rather than the recording.
This required the separation out of the code which calculates the statistics from the code which converts the recording into a spectrogram.
Have added another statistical property, a count of the number of spectral peaks in the spectral profile of the event.